Change option-parsing strategies
Use `lexopt`, a lower level crate, instead of `clap`. Specifically
`wasm-ld` is a `ld`-like linker and has options that clap doesn't
support:

* Flags like `-shared`
* Positional flags like `--whole-archive` and `--no-whole-archive`

To handle this the `lexopt` crate is much lower level. It means we don't
get nice help messages but an effort to add that back in can be added
later as well.